Implement hangman gameplay
The Game class executes the game with the run method which
loads the dictionary, select a random word and goes through
the gameplay loop. The play method is responsible for the
loop, initializing the gameplay variables, checking for
endgame conditions, requesting input from the player and
updating the variables based on the input. Lastly, the play
method shows the result after the loop exits.